Update 1-29-18
Reformated ERROR handeling with a new Error struct implementing the Error() function from error interface.

Removed old help code.  Started implementing a basic !help structure with in the help_commands.go file

atemtped to implement a Gob encoding function to save the maps to a file.  Currently not functioning and commented out.  Requiers modifications and testing to fix.

added guild_command.go for posibal guild commands.  To be implemented later.

Implemented an error check for trying to deleted more then 99 messages at one time.  BulkDelete with in discord only supports up to 99 messages at one time so an error is thrown if this value is exeded.

Added an error check for if you do not give a time to change the ban time to under !player -bantime <time> command.

storage.go holds all functions related to the encoding and decoding of Gob.  Current implementation may need to be modified.

ToLower() updated to return an error if the given location is out of bounds of the array.

utility_commands.go wrapper and shell implemented.  Needs lower functions to be implemented.