Update 2-2-18
Fixed an error in the Note on the README.

Started implementation of the !help command structure.  Still needs a lot of work and is only a basic shell.

Fixed Error creation as there was a mistake in not created the Error struct correctly.

Moved save and load to bot.go for clarity.  It didnt make sense to have it with in command...

Added -clear or -cl to !message.  This clears messages in the server up to 99 at a time.  There is a limit to deleting 99 messages at a time.  May change to delete all messages in pageges of 99 till it returns no messages to delete.

Removed Loging from util.go as it know has error handeling and should not be printing from its own function.