Update 2-6-18 2
Changed date and time format of errors to be more readibal.  Know fallows normal Unix formating.

Updated main.go and bot.go to log info on the startup of the but.  May be removed later but implace for testing perpuses.

Updated README to be more accuret.  Command charts are know up to date and give accurent information on the current state of the bot.

Removed the instalation instroctions as the normal user will not need to install the bot on there computer.  The bot will be running on a AWS server when it is fully deploid.

Link to add the bot to your server is left as is and changed to be the only step requiered to add the bot to your server.

 Noted that the Initial Setup section of the README is still out of date and will need to be changed at a later date.

The info on the Interoducftion could also use to be updated.