Update 2-9-18
Updated MesssageCreates BotCommands map call to be more stream lined.  Removed the redundent map call after the ok check and merged the ok check with the function call.

Reworked the sub commands functionality to use maps insted of switch case statements.  These keep the same relitive format as BotCommands.

Because of the sub function change, player_commands.go had to go through major format changes.  A lot of function redundency was eliminated and the function call structure was hevily streamlined.

Renamed LoadBotCommands() to LoadMaps() for clarity.

moved HelpMap out of help_commands.go and in to commands.go with the rest of the maps

Moved some functions to private as they they had no need to be exported.