Update 2-9-18 2
message_command.go function optimization.  Reorginized the delete functions to be a little simpler and optimized.   Removed unneeded if checks and switch from an if case check of length to a switch case check.

Start of reformating error checks with in code.  Changing error if check to one line rather then two for redability.

Reworked bantime to only need to call the map once when running -ban.

Changed -ban to check if the bantime as be set for that server.  If it has not been set for what ever reson it will through an error to the chat telling you to set the bantime.

Changed -bantime to check if it has been set on the server before trying to desplay the prior ban time before changing it to the new ban time.

Made banTime map a non-exported map as it should not be exported outside of the package.