Update 2-9-18
Finished refactoring code to handle errors with the new style.

Fixed a few bugs that were found well refactoring code.

Fully updated all comments with in the code.  All comments with in all files for all functions that are currently implemented are up to date.

Exceptions are filter.go and storage.go as they are still being worked on.

Added error handaling on CommanD-Bot.New().

Removed loging when bot started as it was anoing and only there for the fun of it well testing.

Other small code cleenups and format changes.