Update 2/17/18
Changed ToString() to take a seporator string which lets me pick which char I want to seporate on.

Large scale refatoring of map structure.  BotCommands map know maps to the sub maps of the corisponding sub functions.  This eliminated a good amount of reepet code and a hole function from each command file.

moved admin check to the head of each function as a way to simplify and unify function declorations.

First pass at fixing an issue were !ms -cl would not work when a message with in a server was over 14 days old.  Current version is not functioning.

Implimented a check in to ToDelete which checks the yera and day of the messages creation so that it will not atempt to delete messages older then 14 days.

Due to repeted admin checks in each function witht he new system, IsAdmin was added to server.go to prevent code repotition.

botSession in main.go changed to fallow the same error checking format as the rest of the code.