Update 3/19/18
Code refactoring and such as always.

- Redid ToLower in to two functions:
*ToLower() []string
*StrToLower()string
this is to get rid of needing to pass a index in to ToLower so that I can set the hole string to lower case if I want to.  ToLower is just a wraper for StrToLower which loops through the passed in string array and set it to lower and append it to a new array of strings to return.  Removed error checking with in function as it is no longer needed.

added channel_commands.  Not fully tested currently so there may be errors.  Testing and refining is part of TODOs

Changed BanTime to a map of key string to int for simplisity.  There was no need for the cmdInfo struct.  Also made it private to commands as it should not be accessed outside of that package.

removed !help for future modifictation.  !help as always needs a full overhall again as it was geting out of date and needs a better, more perminit salution rather then the hardcoded version i was using prior.

Added IGN and Trinity for basic testing.  Needs full database implimentation and saving as all data is wiped on restart.

Made ToDelete exported as it is needed outside of the command package for the filter.  May need some minor refactoring or a wrapper for filter simplisity.

 Added error handling to !ms -cl with in loop.

Added utility_commands diceRole.  Untested but basic functionality is currently there.

Split ParceInput and Parce up.  ParceInput is a wrapper of Parce to split a string on a space.  Parce splits a string on any given deliniator.

Added IsTime in to utility for time message time checks.  This is to prevent errors when trying to delete messages that are older then 14 days.  Prior you could not call !ms -cl on a list of message with any messages over 14 days as it would just error out and not delete any.  Now if it hits a message that is 14 days or older it returns the current list of messages to delete (if any) and trys to delete them.  It will then end as there are no more messages to delete.

Implimentation of filter.  The filter is still very simple and needs a lot of work but it currently funtions on some basic sware words and such.